Add files via upload
Business problem Identification

In any business, it is important to get the customer feedback on a product. Accumulating feedback from multiple sources allows for further product improvement or marketing adjustments. Creating a feedback analysis model that can reliably assess customer reviews for car products is the challenge. Feedback can be categorized using such a model into positive, neutral, and negative categories. The reviews that are provided indicate that the focus is specifically on comprehending customers' experiences with car dusters and booster cables.